Fix double-image bug
There was a bunch of complex logic to avoid both reloading an image after it got uploaded to S3 while also not storing the raw base64 image data in localstorage. The complexity of having the localStorage state be different from the Rect/view model state made it difficult to reason about, and the result was having previous images show up twice.

I ripped out all that logic, and it seems to have fixed the double-image problem.